BankApp_Frontend/requirements.txt
flask
flask-cors
pymongo
bcrypt
PyJWT
